requirements.txt
python-telegram-bot==21.1
matplotlib
pandas



